tools/bitbucket/requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

annotated-types==0.7.0
anyio==4.9.0
atlassian-python-api[bitbucket]==4.0.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1; platform_python_implementation == "CPython" and sys_platform == "win32"
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6; platform_system == "Windows"
deprecated==1.2.18
dify_plugin>=0.1.0,<0.2.0
dpkt==1.9.8
flask==3.0.3
gevent==24.11.1
greenlet==3.2.1; platform_python_implementation == "CPython"
h11==0.14.0
httpcore==1.0.8
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
markupsafe==3.0.2
multidict==6.4.3
oauthlib==3.2.2
propcache==0.3.1
pycparser==2.22; platform_python_implementation == "CPython" and sys_platform == "win32"
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pydub==0.25.1
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==79.0.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
tiktoken==0.8.0
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
werkzeug==3.0.6
wrapt==1.17.2
yarl==1.20.0
zope-event==5.0
zope-interface==7.2
